refactor: clarify auth and settings layers
