+++ src/shared/settingsView/toolDashboardData.ts
